Commit to the initial tests
We have a little base that returns empty HTML files,
provides a favicon.png and has a couple of little tests
that describe how the API should end up looking like. The
views will get built on top of those returned values, either
on the server side or as a single page application on the
client's end.