Restructure the data we return for a Person
Morning, Sunday! Hope everyone's had a great Saturday.

This change adjusts our data structures to work with the data
that is quickly returned to us by TMDb inside of a single request
since we do not want to be making multiple requests and should be
able to work with what we have there. :+1: