Partially test the image URLs
We should not hardcode the URLs and check whether they match since the
domain/path can change at any point (there is a reason why it is
provided as a dynamic configuration). Instead, we can test whether it
starts off with a `https://` so that we do not load anything over plain
HTTP ever and ends with a `.jpg` since that appears to be the file type
in use and I do not see them switching away any time soon then.