Add Bower for front-end libraries
We should be able to work with the small backend that we have laid
out here. Time to move on to the front-end part of this. Might
return to caching those TMDb queries later down the line - still
got that Redis dependency laying around there - but need to get
around to that. First, rough front-end.