Assist the CI builder
Travis is complaining that C++ for native library building is not
available. Right now we aren't really building anything, but should
probably set that up for future-proofing, if we have the choise of
building on the Trusty images instead of the Precise images.

Additionally, the `jshint` call has been dropped as we are not going
to be including that package since it has specified in its package.json
that it prefers to be installed globally and we do not want warnings
based on that. We still have `semistandard` as our linter and it is
a pretty good one.