Add a movie detail listing front-end
Yes, it may not look very good, but it is pretty damn functioning.
Just need to link this up with the actor details and we are all set.
Well, yeah, possibly after a pair of hours sleep and a sleepy
bus ride.

Tired developers do not make good decisions.
Don't forget sleep is important.