Add Gulp for dealing with building
Now we have a /build/ instead of a /static/. It gets generated on
the go. The LESS files are automatically made into CSS and compressed.
JavaScript is compressed as well. Same thing with images. Oh, and
HTML templates are built right away and served as proper static files.
We are not using Pug in Express for rendering anyway, so no need to
keep it there.

Still need to make this thing watch file changes and possibly run
them tests upon changes.