Add Mocha to Gulp
We want tests to be run interactively by default.

This leaves us with the major tasks of `build` and `mocha`
along with their `watch-*` counterparts. The `default` now
just runs all the things. That also means that for the
default invocation, the API key for TMDb must be set since
our tests query the servers, but hey, should not be an issue.