Do naive caching on queries against movies
We can just cache the whole responses for a short period of time so
that we can return later responses much much quicker. It is probably
a common occurance that people return to the same movie searches and
that there are some queries that will be popular regardless and can
benefit from global caching.